Refactor ImageGallery component: separate client logic into ImageGalleryClient for improved modularity and maintainability. Introduce ImageGalleryClientProps type for better prop management.
